move get_llm_from_env
