+++ bot/utils.py
@@ -1,6 +1,4 @@
 import contextlib
-import functools
-import os
 import re
 import tempfile
 from pathlib import Path
@@ -12,7 +10,6 @@ from langchain_community.document_loaders import YoutubeLoader
 from langchain_community.document_loaders.html_bs import BSHTMLLoader
 from langchain_community.document_loaders.pdf import PyPDFLoader
 from langchain_core.documents import Document
-from langchain_core.language_models.chat_models import BaseChatModel
 from langchain_core.messages import AIMessage
 from loguru import logger
 
@@ -69,20 +66,6 @@ def set_sqlite_llm_cache() -> None:
     cache = SQLiteCache(database_path.as_posix())
 
     set_llm_cache(cache)
-
-
-@functools.cache
-def get_llm_from_env() -> BaseChatModel:
-    if "OPENAI_API_KEY" in os.environ:
-        from langchain_openai import ChatOpenAI
-
-        return ChatOpenAI(model="gpt-4o-mini", temperature=0)
-    elif "GOOGLE_API_KEY" in os.environ:
-        from langchain_google_genai import ChatGoogleGenerativeAI
-
-        return ChatGoogleGenerativeAI(model="gemini-1.5-flash", temperature=0)
-    else:
-        raise ValueError("No API key found in environment variables")
 
 
 def ai_message_repr(ai_message: AIMessage) -> str:
